requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==24.1.0
pyyaml==6.0.1
python-dateutil==2.8.2
diskcache==5.6.3
pytest==7.4.0
pytest-asyncio==0.21.1

# Code quality tools
ruff==0.1.6
black==23.11.0
vulture==2.10
bandit[toml]==1.7.5
pre-commit==3.5.0
